refactor OnGitHubLink code location
